routes: Fix index path
Previously, routes.js used the wrong path to get the index.html file. This didn't
seem to affect the user experience, but the logs would have a bunch of errors.
Since we are going to use this blueprint for the getting started guide, we do
not want all these errors in the logs.